Move timeseries cli entry point to cli.
